+++ app/error-message-status.js
@@ -1,0 +1,5 @@
+module.exports = {
+  DEFAULT: false,
+  BYPASS: 'BYPASS',
+  SKIP: 'SKIP',
+};